Handling players and games is now getting a little complicated.
Previously we had just let character IDs be added directly to games through the game runner.  This worked, but
now that I am wanting to allow the game runner to issue messages to players who are affected by activities
in the game, this model is starting to break down.

What we need is the ability to relate players to games, so that when a game even occurs that affects one or more player's characters,
the player can be sent a message.  Sending messages can be accomplished by using bog standard tokio channels easily enough, and
we don't even need to mix in web metaphors to do it so we get nice clean boundaries.  However, we now need to do extra work when
adding a player to a game, or deleting a game with players still active etc.  It's probably best to move this functionality into
its own box - the registry.

The registry isn't done yet; it doesn't have much work to do, though, so I expect it won't take long to finish.  It can currently
accept player IDs and MPSC senders, return a sender on demand, associate games with game IDs, return mutable references to games
and associate players and games together.

We still want to have proper clean up - that is, when a game is destroyed all of the players should be unlinked.  Probably a few
other things will occur to me as we go.  Once it's mostly done, though, we can strip out the current ad hoc hashmap in the runner
and replace it with the registry.